添加授权

+++ RazorPagesPizza/Pages/Pizza.cshtml.cs
@@ -2,11 +2,16 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using RazorPagesPizza.Models;
 using RazorPagesPizza.Services;
+using Microsoft.AspNetCore.Authentication;
+using Microsoft.AspNetCore.Authorization;
 
 namespace RazorPagesPizza.Pages
 {
+    [Authorize]
     public class PizzaModel : PageModel
     {
+        public bool IsAdmin => HttpContext.User.HasClaim("IsAdmin", bool.TrueString);
+
         public List<Pizza> pizzas = new();
 
         [BindProperty]
@@ -17,8 +22,10 @@ namespace RazorPagesPizza.Pages
             pizzas = PizzaService.GetAll();
         }
 
+        
         public IActionResult OnPost()
         {
+            if (!IsAdmin) return Forbid();
             if (!ModelState.IsValid)
             {
                 return Page();
@@ -29,6 +36,7 @@ namespace RazorPagesPizza.Pages
 
         public IActionResult OnPostDelete(int id)
         {
+            if (!IsAdmin) return Forbid();
             PizzaService.Delete(id);
             return RedirectToAction("Get");
         }
